+++ collector/collector.go
@@ -198,3 +198,49 @@ var ErrNoData = errors.New("collector returned no data")
 func IsNoDataError(err error) bool {
 	return err == ErrNoData
 }
+
+// pushMetric helps construct and convert a variety of value types into Prometheus float64 metrics.
+func pushMetric(ch chan<- prometheus.Metric, fieldDesc *prometheus.Desc, name string, value interface{}, valueType prometheus.ValueType, labelValues ...string) {
+	var fVal float64
+	switch val := value.(type) {
+	case uint8:
+		fVal = float64(val)
+	case uint16:
+		fVal = float64(val)
+	case uint32:
+		fVal = float64(val)
+	case uint64:
+		fVal = float64(val)
+	case int64:
+		fVal = float64(val)
+	case *uint8:
+		if val == nil {
+			return
+		}
+		fVal = float64(*val)
+	case *uint16:
+		if val == nil {
+			return
+		}
+		fVal = float64(*val)
+	case *uint32:
+		if val == nil {
+			return
+		}
+		fVal = float64(*val)
+	case *uint64:
+		if val == nil {
+			return
+		}
+		fVal = float64(*val)
+	case *int64:
+		if val == nil {
+			return
+		}
+		fVal = float64(*val)
+	default:
+		return
+	}
+
+	ch <- prometheus.MustNewConstMetric(fieldDesc, valueType, fVal, labelValues...)
+}
