--- collector/utils.go
@@ -1,66 +1,0 @@
-// Copyright 2022 The Prometheus Authors
-// Licensed under the Apache License, Version 2.0 (the "License");
-// you may not use this file except in compliance with the License.
-// You may obtain a copy of the License at
-//
-// http://www.apache.org/licenses/LICENSE-2.0
-//
-// Unless required by applicable law or agreed to in writing, software
-// distributed under the License is distributed on an "AS IS" BASIS,
-// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-// See the License for the specific language governing permissions and
-// limitations under the License.
-
-//go:build !nonetclass && linux
-// +build !nonetclass,linux
-
-package collector
-
-import (
-	"github.com/prometheus/client_golang/prometheus"
-)
-
-func pushMetric(ch chan<- prometheus.Metric, fieldDesc *prometheus.Desc, name string, value interface{}, valueType prometheus.ValueType, labelValues ...string) {
-	var fVal float64
-	switch val := value.(type) {
-	case uint8:
-		fVal = float64(val)
-	case uint16:
-		fVal = float64(val)
-	case uint32:
-		fVal = float64(val)
-	case uint64:
-		fVal = float64(val)
-	case int64:
-		fVal = float64(val)
-	case *uint8:
-		if val == nil {
-			return
-		}
-		fVal = float64(*val)
-	case *uint16:
-		if val == nil {
-			return
-		}
-		fVal = float64(*val)
-	case *uint32:
-		if val == nil {
-			return
-		}
-		fVal = float64(*val)
-	case *uint64:
-		if val == nil {
-			return
-		}
-		fVal = float64(*val)
-	case *int64:
-		if val == nil {
-			return
-		}
-		fVal = float64(*val)
-	default:
-		return
-	}
-
-	ch <- prometheus.MustNewConstMetric(fieldDesc, valueType, fVal, labelValues...)
-}
